requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
torchaudio==2.1.0
numpy<2.0.0
python-dotenv==1.0.0
pydantic==2.5.0
websockets==12.0
aiofiles==23.2.1

